X86: Pass SMP and SMT info in AT_HWCAP2
On x86, we can get number of available logical processors from CPUID.
But some logical processors may be disabled by BIOS or kernel.  To
compute cache sizes shared by logical processors in user spaces, we need
to know if SMP and SMT are enabled.  This patch adds X86_FEATURE_SMP
and X86_FEATURE_SMT to x86_capability[7] to pass it in AT_HWCAP2 to
user space.

Signed-off-by: H.J. Lu <hjl.tools@gmail.com>